refactor: move PostgresRawQuery to _queries module
